:recycle: refactor: move queries from api endpoint to reuseful
