+++ server/db/queries/location.ts
@@ -1,0 +1,54 @@
+import { and, eq } from "drizzle-orm";
+import { customAlphabet } from "nanoid";
+
+import type { InsertLocation } from "../schema";
+
+import db from "..";
+import { location } from "../schema";
+
+const nanoid = customAlphabet("abcdefghijklmnopqrstuvwxyz0123456789", 5);
+
+export async function findLocationByName(existing: InsertLocation, userId: number) {
+  return db.query.location.findFirst({
+    where: and(
+      eq(location.name, existing.name),
+      eq(location.userId, userId),
+    ),
+  });
+}
+
+export async function findLocationBySlug(slug: string) {
+  return db.query.location.findFirst({
+    where: eq(location.slug, slug),
+  });
+}
+
+export async function findUniqueSlug(slug: string): Promise<string> {
+  let existingSlug = !!(await findLocationBySlug(slug));
+
+  while (existingSlug) {
+    const id = nanoid();
+    const idSlug = `${slug}-${id}`;
+    existingSlug = !!(await findLocationBySlug(idSlug));
+    if (!existingSlug) {
+      return idSlug;
+    }
+  }
+
+  // 如果原始 slug 不存在，直接返回
+  return slug;
+}
+
+export async function insertLocation(
+  insertable: InsertLocation,
+  slug: string,
+  userId: number,
+) {
+  const [created] = await db.insert(location).values({
+    ...insertable,
+    userId,
+    slug,
+  }).returning();
+
+  return created;
+}
